Rename json module to helpers
